rename scheduling component
